Fix page number 0 in pdf table of contents
Some of the files were named with capital letters (5-Lawson.md) but
when built through Hugo the URLs are lowercase, as they should be
(5-lawson/index.html). Because of the capitals in the filenames though,
Prince XML wasn’t properly parsing the file location as was inserting
page numbers of 0 (zero) in the TOC. This changes the filenames to
lowercase, which needed to be done with a git move command (`git mv
5-Lawson.md 5-lawson.md`) otherwise the change wasn’t being noted in
the repo.