Add `sort_by` attribute to biblio for fine tuning
There were a half-dozen entries in the biblio that weren't alphabetizing the way the editors wanted. This commit copies the `short` attribute to `sort_by`, points the shortcode to sort by the `sort_by` attribute rather than `short`, and then adjust the values of a half dozen entries in order to alphabetize them properly.